+++ Kernel/System/Ticket/CustomerPermission/CustomerIDCheck.pm
@@ -25,6 +25,9 @@ sub new {
         $Self->{$_} = $Param{$_} || die "Got no $_!";
     }
 
+    # disable output of customer company tickets
+    $Self->{DisableCompanyTickets} = $Self->{ConfigObject}->Get('Ticket::Frontend::CustomerDisableCompanyTicketAccess');
+
     return $Self;
 }
 
@@ -38,6 +41,9 @@ sub Run {
             return;
         }
     }
+
+    # disable output of customer company tickets if configured
+    return if $Self->{DisableCompanyTickets};
 
     # get ticket data
     my %Ticket = $Self->{TicketObject}->TicketGet(