+++ CHANGES.md
@@ -1,4 +1,5 @@
 #3.3.4 201?-??-??
+ - 2013-12-12 Added functionality to disable access to tickets of other customers with the same customer company in customer interface.
  - 2013-12-12 Fixed bug#[9702](http://bugs.otrs.org/show_bug.cgi?id=9702) - Wrong article type with external ticketnumber recognition i.c.w. follow-up reject option.
  - 2013-12-11 Fixed bug#[9723](http://bugs.otrs.org/show_bug.cgi?id=9723) - TicketAccountedTime stat does not run on Oracle with many tickets
  - 2013-12-09 Fixed bug#[10078](http://bugs.otrs.org/show_bug.cgi?id=10078) - Installer: error log after creating mysql database.